requirements.txt
transformers==4.41.2
torch>=2.3.0
sentencepiece           # 
faiss-cpu

openai
langchain
pandas
streamlit
reportlab
transformers>=4.41.0
torch>=2.2.2
python-dotenv>=1.0.1